Add util func for loading fonts in Storybook, export to consumers
