+++ .storybook/config.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { configure, addDecorator } from '@storybook/react';
 import { withA11y } from '@storybook/addon-a11y';
-import { fontUrl } from '../src/components/shared/global';
+import { loadFontsForStorybook } from '../src/utils/index';
 import 'storybook-chromatic';
 
 import { GlobalStyle } from '../src/components/shared/global';
@@ -23,15 +23,4 @@ configure(
   module
 );
 
-// Load the font and avoid re-loading it when components change
-const fontLinkId = 'font-link';
-
-if (!document.getElementById(fontLinkId)) {
-  const fontLink = document.createElement('link');
-
-  fontLink.id = fontLinkId;
-  fontLink.href = fontUrl;
-  fontLink.rel = 'stylesheet';
-
-  document.head.appendChild(fontLink);
-}
+loadFontsForStorybook();
